refactor: extract cli/ module from lib.rs (#112 PR 4)
Move all CLI subcommand handlers from lib.rs into src/cli/:

- cli/policy_test.rs (371L): `omamori test` + PolicyTestResult + run_policy_tests
- cli/config_cmd.rs (561L): config/override/init + mutate_config (DO NOT SPLIT, T6)
- cli/audit_cmd.rs (201L): audit verify/show/key
- cli/install.rs (214L): install + uninstall
- cli/status.rs (149L): status --refresh

lib.rs: 1,534 → 103 lines (97% reduction from original 2,893).

Public API re-exports maintained:
  pub use cli::policy_test::{PolicyTestResult, run_policy_tests};

Tests moved: 4 (policy_tests 2 + GR-005 mutate_config 2).
Total test count: 473 (unchanged).
